rename query & handler
