+++ src/PrestaShopBundle/Resources/config/services/adapter/product/_combination.yml
@@ -104,15 +104,15 @@ services:
             - name: tactician.handler
               command: PrestaShop\PrestaShop\Core\Domain\Product\Combination\Command\SetCombinationSuppliersCommand
 
-    prestashop.adapter.product.combination.query_handler.get_combination_supplier_options_handler:
-        class: PrestaShop\PrestaShop\Adapter\Product\Combination\QueryHandler\GetCombinationSupplierOptionsHandler
+    prestashop.adapter.product.combination.query_handler.get_combination_suppliers_handler:
+        class: PrestaShop\PrestaShop\Adapter\Product\Combination\QueryHandler\GetCombinationSuppliersHandler
         arguments:
             - '@prestashop.adapter.product.repository.product_supplier_repository'
             - '@prestashop.adapter.product.combination.repository.combination_repository'
             - '@prestashop.adapter.product.repository.product_repository'
         tags:
             - name: tactician.handler
-              command: PrestaShop\PrestaShop\Core\Domain\Product\Combination\Query\GetCombinationSupplierOptions
+              command: PrestaShop\PrestaShop\Core\Domain\Product\Combination\Query\GetCombinationSuppliers
 
     prestashop.adapter.product.combination.command_handler.remove_all_associated_combination_suppliers_handler:
         class: PrestaShop\PrestaShop\Adapter\Product\Combination\CommandHandler\RemoveAllAssociatedCombinationSuppliersHandler
